Some more transaction handling cleanup.
